Did some Visual Studio code analysis and fixed teambuilder

+++ LeagueClient/Logic/com/riotgames/platform/summoner/Summoner.cs
@@ -19,7 +19,7 @@ namespace LeagueClient.Logic.Riot.Platform {
     public Boolean HelpFlag { get; set; }
 
     [SerializedName("sumId")]
-    public Double SumId { get; set; }
+    public Int64 SumId { get; set; }
 
     [SerializedName("profileIconId")]
     public Int32 ProfileIconId { get; set; }
